+++ src/test/resources/design-docs/example/filters/example_filter.js
@@ -1,7 +1,0 @@
-function(doc, req) {
-  if (doc.title == req.query.somekey1) {
-    return true;
-  } else {
-    return false;
-  }
-}