+++ src/test/resources/design-docs/example/rewrites/rewrites.js
@@ -1,8 +1,0 @@
-[
-    {
-        "from": "",
-        "to": "index.html",
-        "method": "GET",
-        "query": {}
-    }
-]