+++ src/test/resources/design-docs/example/validate_doc_update/validate_doc_update.js
@@ -1,1 +1,0 @@
-function(newDoc, oldDoc, userCtx) {}