+++ src/test/resources/design-docs/example/views/boolean/map.js
@@ -1,3 +1,0 @@
-function(doc) {
-  emit(doc.contentArray[0].boolean, null);
-}