+++ src/test/resources/design-docs/example/views/by_tag/map.js
@@ -1,7 +1,0 @@
-function(doc){
-  if(doc.title && doc.tags){
-    for(var idx in doc.tags){
-      emit(doc.tags[idx], 1);
-    }
-  }
-}