+++ src/test/resources/design-docs/example/views/creator_boolean_total/map.js
@@ -1,4 +1,0 @@
-function(doc) {
- emit([doc.contentArray[0].creator, doc.contentArray[0].boolean,
- doc.contentArray[0].created], [doc.title, doc.contentArray[0].total]);
-}