+++ src/test/resources/design-docs/example/views/quotes_created/map.js
@@ -1,3 +1,0 @@
-function(doc) {
- emit([doc.contentArray[0].quotes, doc.contentArray[0].created], null);
-}