+++ src/test/resources/design-docs/example/views/spaces_created/map.js
@@ -1,3 +1,0 @@
-function(doc) {
- emit([doc.contentArray[0].spaces, doc.contentArray[0].created], null);
-}