+++ src/test/resources/design-docs/views101/views/diet_sum/map.js
@@ -1,5 +1,0 @@
-function(doc) {
-	if(doc.diet) {
-		emit(doc.diet, 1);
-	}
-}