+++ src/test/resources/design-docs/views101/views/latin_name/map.js
@@ -1,5 +1,0 @@
-function(doc) { 
-	if(doc.latin_name){
-		emit(doc.latin_name, doc.latin_name.length);
-	}
-}