+++ src/test/resources/design-files/conflicts_design_doc.js
@@ -1,0 +1,9 @@
+{
+  "_id": "_design/conflicts",
+  "language": "javascript",
+  "views": {
+    "conflict": {
+      "map": "function(doc){ if(doc._conflicts) emit(doc._conflicts, null);}\n"
+    }
+  }
+}